// generated/src/app/api-models/tooth-history.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ToothHistory {
  seqToothHistoryId: number;
  seqMembId: number;
  dateOfService: Date;
  toothNumber: string;
  procedureCode: string;
  surface1: string;
  surface2: string;
  surface3: string;
  surface4: string;
  surface5: string;
  arch: string;
  quadrant: string;
  oralCavity: string;
  toothStatus: string;
  seqProvId: number;
  benefitPackageId: string;
  seqClaimId: number;
  lineNumber: number;
  subLineCode: string;
  sourceType: string;
  userDefined1: string;
  userDate1: Date;
  userDefined2: string;
  userDate2: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  medDefCode: string;
}